# server/docs/swagger.yaml
basePath: /v1
definitions:
...
  handler.CreateItemReq:
    properties:
      category_id:
        type: integer
      condition:
        type: string
      description:
        type: string
      estimated_value:
        type: number
      images:
        items:
          type: string
        type: array
      location:
        type: string
      title:
        type: string
      video_url:
        type: string
      want_items:
        items:
          type: string
        type: array
    required:
    - condition
    - title
    type: object
...
  handler.LoginReq:
    properties:
      code:
        type: string
      phone:
        type: string
      user_type:
        description: personal / enterprise
        type: string
    required:
    - code
    - phone
    type: object
...
  handler.SendCodeReq:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
...
  handler.UpdateStatusReq:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  handler.UpdateUserReq:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      location:
        type: string
      nickname:
        type: string
    type: object
  model.Category:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
    type: object
...
  model.Item:
    properties:
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: integer
      condition:
        type: string
      created_at:
        type: string
      description:
        type: string
      estimated_value:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      location:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 关联
      user_id:
        type: integer
      video_url:
        type: string
      view_count:
        type: integer
      want_items:
        items:
          type: string
        type: array
    type: object
...
  model.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      credit_score:
        type: integer
      enterprise_name:
        type: string
      enterprise_verified:
        type: boolean
      id:
        type: integer
      last_login_at:
        type: string
      location:
        type: string
      nickname:
        type: string
      phone:
        type: string
      real_name_verified:
        type: boolean
      status:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: NeoBarter Team
    url: https://github.com/sxc19960907/neobarter
  description: 现代易货交易平台后端 API 文档。支持 C2C/B2B 双模式、巴特币结算、智能匹配。
  license:
    name: MIT
  termsOfService: https://github.com/sxc19960907/neobarter
  title: 新易巴特 NeoBarter API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 手机号+验证码登录，未注册的手机号自动创建账户并赠送初始巴特币
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回 token 和用户信息
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 登录或注册
      tags:
      - 认证
  /auth/send-code:
    post:
      consumes:
      - application/json
      description: 向指定手机号发送登录/注册验证码（60秒内不可重复发送）
      parameters:
      - description: 手机号
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SendCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送验证码
      tags:
      - 认证
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Category'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取分类列表
      tags:
      - 分类
  /items:
    get:
      description: 支持分类、关键词、地区、成色、估值范围筛选；mine=true 查询自己的物品
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 分类ID
        in: query
        name: category_id
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 地区
        in: query
        name: location
        type: string
      - description: 成色
        in: query
        name: condition
        type: string
      - description: 最低估值
        in: query
        name: min_value
        type: number
      - description: 最高估值
        in: query
        name: max_value
        type: number
      - description: 排序：created_at/view_count/estimated_value
        in: query
        name: sort_by
        type: string
      - description: 是否只看自己的物品
        in: query
        name: mine
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 获取物品列表
      tags:
      - 物品
    post:
      consumes:
      - application/json
      description: 发布一个新的可交换物品
      parameters:
      - description: 物品信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateItemReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Item'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 发布物品
      tags:
      - 物品
  /items/{id}:
    delete:
      parameters:
      - description: 物品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除物品
      tags:
      - 物品
    get:
      parameters:
      - description: 物品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Item'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取物品详情
      tags:
      - 物品
    put:
      consumes:
      - application/json
      parameters:
      - description: 物品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 物品信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateItemReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Item'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新物品信息
      tags:
      - 物品
  /items/{id}/status:
    put:
      consumes:
      - application/json
      description: active=上架, inactive=下架
      parameters:
      - description: 物品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改物品状态
      tags:
      - 物品
...
  /users/{id}:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户公开信息
      tags:
      - 用户
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 用户
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新当前用户信息
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    description: 输入格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"